backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
gunicorn==21.2.0
PyPDF2==3.0.1